fix(samples): move new samples out of samples and topics subtrees so synthtool can find them
