+++ samples/deleteTopic.js
