+++ samples/getTopicPolicy.js
