+++ samples/listSubscriptions.js
