+++ samples/listenForErrors.js
