+++ samples/publishBatchedMessages.js
