+++ samples/publishMessage.js
