+++ samples/publishMessageWithCustomAttributes.js
