+++ samples/setTopicPolicy.js
