+++ samples/setup_for_tests.sh
@@ -1,0 +1,3 @@
+export GOOGLE_APPLICATION_CREDENTIALS=/Users/mzp/git/keys/google-cloud-node-ca5ce18b914a.json
+gcloud config set project long-door-651
+export GCLOUD_PROJECT=long-door-651
