+++ samples/synchronousPull.js
