+++ samples/testSubscriptionPermissions.js
