+++ samples/testTopicPermissions.js
